update zigbee with client/server
